Update game to the end of episode 43
In this update, cannons have been updated so that
they track the first enemy in the leading wave
currently on screen (or the next one that appears).

A splash screen has been added to the beginning
of the game. The user can select from PLAY, EDITOR,
and QUIT. PLAY opens the hardcoded grid and spawns
enemies immediately. EDITOR opens to the last
saved map the user created (or a completely grass-
filled grid if no map has been saved). The user can
create a custom map and save the map to edit again
later. QUIT closes the game.

Finally, projectiles know when they have collided
with their target and when the collision happens,
the projectile is no longer drawn on screen. It is
still present in the list of projectiles for the
cannon that fired it.

+++ src/helpers/Clock.java
@@ -43,7 +43,7 @@ public class Clock {
 		totalTime += d;
 	}
 	
-	public static void ChangeMultiplier(int change) {
+	public static void ChangeMultiplier(float change) {
 		if (multiplier + change < -1 || multiplier + change > 7) {
 			
 		} else {
